Change "email" to "message" in event name.
The event handles sending texts through Message Broker as well, so
email naming was confusing.
